Simple models supporting one country
Has relationship between Plan and Benefit that I think it good but,
I'm thinking I should extract out country.  I wasn't going to
initially but, the fact that I have two pieces of data (country code
and currency) in the Plan that are related makes me think it should be
broken out into a different model.